Begin Milstone component rename to Release
